requirements.txt
fastapi==0.115.11
firebase_admin==6.6.0
aiohttp==3.9.5
git+https://github.com/ozgur/python-firebase
google-generativeai
gspread==5.12.4
google-auth==2.35.0
line-bot-sdk==3.11.0
Pillow==11.1.0
pydantic==2.10.6
tiktoken==0.9.0
uvicorn[standard]==0.34.0
# yfinance